refactor: mv order margin method to sep file
