+++ tests/unit/kite/market.cpp
@@ -304,56 +304,3 @@ TEST(kiteTest, getInstrumentsExchangeTest) {
     EXPECT_EQ(instrument2.segment, "NSE");
     EXPECT_EQ(instrument2.exchange, "NSE");
 }
-
-TEST(kiteTest, getOrderMarginsTest) {
-    const string JSON =
-        kc::test::readFile("../tests/mock_responses/order_margins.json");
-    constexpr double QUANTITY = 1;
-    constexpr double PRICE = 0;
-    constexpr double TRIGGER_PRICE = 0;
-    const string EXCHANGE = "NSE";
-    const string TRADINGSYMBOL = "INFY";
-    const string TRANSACTION_TYPE = "BUY";
-    const string VARIETY = "regular";
-    const string PRODUCT = "CNC";
-    const string ORDER_TYPE = "MARKET";
-
-    StrictMock<kc::test::mockKite> Kite;
-    EXPECT_CALL(Kite,
-        sendReq(utils::http::endpoint { utils::http::METHOD::POST,
-                    "/margins/orders", utils::http::CONTENT_TYPE::JSON },
-            utils::http::Params { { "",
-                R"([{"exchange":"NSE","tradingsymbol":"INFY","transaction_type":"BUY","variety":"regular","product":"CNC","order_type":"MARKET","quantity":1.0,"price":0.0,"trigger_price":0.0}])" } },
-            utils::FmtArgs {}))
-        .Times(1)
-        .WillOnce(
-            Return(ByMove(utils::http::response(utils::http::code::OK, JSON))));
-
-    const std::vector<kc::orderMargins> MARGINS =
-        Kite.getOrderMargins({ kc::orderMarginsParams()
-                                   .Quantity(QUANTITY)
-                                   .Price(PRICE)
-                                   .TriggerPrice(TRIGGER_PRICE)
-                                   .Exchange(EXCHANGE)
-                                   .Tradingsymbol(TRADINGSYMBOL)
-                                   .TransactionType(TRANSACTION_TYPE)
-                                   .Variety(VARIETY)
-                                   .Product(PRODUCT)
-                                   .OrderType(ORDER_TYPE) });
-
-    ASSERT_EQ(MARGINS.size(), 1);
-    kc::orderMargins ordMargins1 = MARGINS[0];
-    EXPECT_EQ(ordMargins1.type, "equity");
-    EXPECT_EQ(ordMargins1.tradingSymbol, "INFY");
-    EXPECT_EQ(ordMargins1.exchange, "NSE");
-    EXPECT_DOUBLE_EQ(ordMargins1.span, 0);
-    EXPECT_DOUBLE_EQ(ordMargins1.exposure, 0);
-    EXPECT_DOUBLE_EQ(ordMargins1.optionPremium, 0);
-    EXPECT_DOUBLE_EQ(ordMargins1.additional, 0);
-    EXPECT_DOUBLE_EQ(ordMargins1.bo, 0);
-    EXPECT_DOUBLE_EQ(ordMargins1.cash, 0);
-    EXPECT_DOUBLE_EQ(ordMargins1.var, 1498);
-    EXPECT_DOUBLE_EQ(ordMargins1.pnl.realised, 0);
-    EXPECT_DOUBLE_EQ(ordMargins1.pnl.unrealised, 0);
-    EXPECT_DOUBLE_EQ(ordMargins1.total, 1498);
-}
